fix(android): move domaindata to dom node
